sidebar-thumbnails: display a first blank thumbnail
when following conditions are met:
 - We are in dual page mode
 - Option for "odd pages on left side" is off
 - We are using the GtkIconView (TODO: GtkTreeView implementation)
 - GtkIconView is displaying two columns of thumbnails

This blank thumbnail can't be selected, and we make sure
the page count logic (when relying on number of items in IconView)
is updated accordingly when we are displaying the blank thumbnail.

The blank thumbnail item can be distinguished among all other items
in the GtkIconView as it's the only one which has the COLUMN_SURFACE
as NULL while COLUMN_THUMBNAIL_SET is set to TRUE.

Addresses issue #30, however there is a corner case still missing
to considerit fixed: When resizing the sidebar, it should keep
the blank thumbnail.