zoom-action: Respond to icon touch events
The zoom entry is filtering out touch events because it assumes that it
will only handle button events. It assumes it will only receive
GdkEventButton events and ignores the touch event because the miscast
data structure doesn't look like it's for the primary mouse button. Fix
the function signature and only filter by the button type if this is a
button event.

Fixes #976