meson: Fix GtkDoc documentation generation
meosn has recently acquired support for `module_version`[0] that
affects the installed location and the devhelp2 file location.

The lack of support for this feature in meson means that the
installation was not the same in meson and in autotools.

GtkDoc documentation has been changed to also use this feature,
producing an equivalent installation.

meson version has been also bumped to 0.49.

[0] https://github.com/mesonbuild/meson/commit/f0c4d1cd00da1f6e7d0126a156928e242d6c5b8b