ci: Run flatpak CI only for master
The current flatpak scripts runs the job using the manifest,
which retrieves and builds the source from git master. Therefore,
it is unhelpful to build the MR.

For the moment, we will build the flatpak module only for
commits on master. Once we fix the script, we can extend the
CI tests to run for MR as well.