ev-view.c: keep cursor visible when changing pages
When zoomed in in non-continuous mode, the caret cursor is sometimes not
visible after moving it to the next or previous page with the up and
down arrow keys.

Call `_ev_view_ensure_rectangle_is_visible` when changing to the next
page, not only when changing to the previous page. Adjust `rect` for
the coordinate difference between the window and the view's scrolling
position.