libview: Handle adjustment change during drag event
If you zoom in or out while using middle-click drag to scroll the
document, evince may unexpectedly jump to a different page. This is
because the zoom operation changes the adjustment's `upper` and `value`
to account for the document's new size while keeping the visible part of
the document visible, but the drag event will overwrite the adjustment
value next time the mouse is moved.

Fix this by updating the drag start position so that the drag can
continue from the new location.