ev-view: fix appearance of editable GtkComboBox in PDF form
Due to a bug in GtkEntry, the GtkEntry of an editable GtkComboBox
will be too wide (due to a hardcoded minimum-width of
150px), overcome this by forcing a very small minimum-width via
"width-chars" property, this allows the GtkEntry to match the
width of the ComboBox item list.

We also tweak evince.css to reduce the padding of the GtkEntry and
GtkButton of the GtkComboBox, which was 8px and made the GtkButton
take up too much space and GtkEntry be left with too little space
for displaying text.

Workarounds gtk#1422

Fixes evince#1002