toolbar: Provide a visual cue when find is unsupported
When document type does not support 'find', Evince disables the
find feature. Previously, we added a tooltip explaining the reason.
However, a straightforward visual cue is more desirable, especially
for touchpad users.

Aside the tooltip, we now also provide an icon that mimic other
icons in GNOME when a service is unavailable, that is, an 'x' on
the icon for mute or wifi.

Fixes #105