meson: don't call cc.has_function() with internal dependencies
When passing internal dependencies in the 'dependencies' parameter of
has_function, meson fails with an error like the following:

  meson.build:165:9: ERROR:  Dependencies must be external dependencies

cc.has_function() cannot be used if the dependency comes from
a subproject; this is because it's not possible to link against an
internal dependency as it has not been built yet at configure time.

NOTE: the current changes only cover the invocations of has_function()
which can actually have internal dependecies (subprojects) passed as
argument.