previewer: Fix compilation on non-unix systems
Compilation on non-unix systems fails with the following message:

../previewer/ev-previewer-window.c:275:13: error: ‘ev_previewer_window_print’ undeclared here (not in a function); did you mean ‘ev_previewer_window_init’?
  { "print", ev_previewer_window_print },
             ^~~~~~~~~~~~~~~~~~~~~~~~~
             ev_previewer_window_init

This build regression on non-unix systems was introduced by commit
65bfd21c (previewer: Use GtkHeaderBar in print preview, 2018-11-08).

During the transition from GtkActionEntry to GActionEntry a check for
GTKUNIXPRINT_ENABLED was forgotten.

Add the check back to make sure that ev_previewer_window_print is used
only if it was defined before.