libview: have EvView use EvDocumentFindMatch
The new find interface returns a GPtrArray of EvDocumentFindMatch,
instead of a GList of EvRectangle.

With EvDocumentFindMatch, each search result has a list of rectangles
instead of a single rectangle. This changes the highlight box drawing to
support multiple rectangles for each search result.

This removes the deprecated function `ev_view_find_changed`, since it
would be difficult to keep that working with the new find interface. The
`ev_view_find_started` function should be used instead.