djvu: use find_text_offset interface for search
This allows the sidebar to show search results for DjVu documents.

The current method of showing results in the sidebar relies on doing a
mapping from the search result's area on the page to the document's
text. This does not work for DjVu documents because they often do not
specify the exact location of characters beyond the location of a line
or word. With the find_text_offset interface, the backend returns an
EvDocumentFindMatch struct, which contains both the area and the offset
into the document's text, so this mapping is not needed.

EvDocumentFindMatch also allows an improvement for multi-line search
results. The current interface only returns a single rectangle for each
result, so for a multi-line search, the rectangle is expanded to cover a
larger area, the bounding box of whole search result.
EvDocumentFindMatch has a list of rectangles, so we can return several
rectangles instead.