shell: prevent multiple fullscreen toolbar timeouts
Gtk is generating a 'leave-notify-event' whenever the fullscreen toolbar
is clicked without a matching 'enter-notify-event', and we create a new
timeout each time. When we exit fullscreen mode, we only clear the last
timeout, so we can end up with a timeout on a null toolbar, leading to a
crash.

Fix this by clearing any existing timeouts before creating a new one.

Fixes #787