ev-page-action-widget: disconnect notify::document signal in finalize
EvPageActionWidget listens for the notify::document signal, but does not
disconnect the signal in finalize, so if you enter and exit fullscreen
mode, which creates and destroys a fullscreen toolbar, then reload the
document, this results in a calling the signal handler with an invalid
EvPageActionWidget pointer.

Disconnect the notify::document signal when EvPageActionWidget is
finalized.

Fixes #789