require Poppler 0.76.0 for ignoreDiacritics feature
instead of 0.73.0 (where it was first implemented)
because 0.76.0 includes a fix for a regression that
caused non-ascii search terms (like in eg. Russian)
to not match anything when using the ignoreDiacritics
option.

The workaround is to use the 'case sensitive' search
option (because using 'case sensitive' automatically
disables ignoreDiacritics inside Poppler).

But given that Fedora 30 (and probably other distros)
ship Poppler 0.73.0, we better require the regression
fixed version of Poppler, so to not regress on users
which search text in the affected languages.

Issue #1179