libview: don't enable popup actions on page change
Action states are reset on page change, so if the page changes while a
popup menu is shown, such as during kinetic scrolling, the popup will
show all items, including the context-sensitive ones which should be
hidden.

Fix this by not enabling popup actions on page change. The appropriate
actions are already enabled when the popup is shown.

Fixes #1168