sidebar links: fix popup
The outline's popup is currently empty because it's being populated with
a label instead of a menu item, and on Wayland, it prints warning
messages because the popup is shown without being attached to a widget.

Change this to use a menu item, and attach the menu to the outline's
tree view.