djvu: casefold search text in case insensitive search
In case-insensitive search, the DjVu backend fails to find any matches
if the search string contains capital letters. This is because the
backend 'casefolds' the document text, converting it to a form that can
be compared independent of case, and which is usually composed of lower
case letters, but doesn't casefold the search text.

Fix this by also casefolding the search text during case-insensitive
search.

Fixes #1212