ev-view-presentation: Don't move goto window offscreen
The gdk_screen_get_width() and gdk_screen_get_height() functions are
deprecated since gdk 3.22. They are used by EvViewPresentation's
ev_view_presentation_goto_window_send_key_event() function to move a
window offscreen.

This function's purpose is to forward an event to the goto window, but
since windows can only receive events when they are 'shown', it first
moves the window offscreen so it won't be visible, then shows the
window, forwards the event, and hides the window again.

Rearrange the code so that the goto window is already visible and in the
correct location before forwarding the event. Remove the now unneeded
showing and hiding code and remove use of the deprecated GdkScreen
functions.

Issue #1023