ev-view-presentation: set goto window's parent
The presentation view has a feature where if you type numbers in
presentation mode, a text entry popup appears so you can enter a page
number to navigate to. The popup does not correctly appear when running
under Wayland and the console shows this message:

"Window 0x562684b01210 is a temporary window without parent, application
will not be able to position it on screen."

Fix this by using the gtk_window_set_transient_for() function to set the
popup's parent.