shell: Show annotation contents in sidebar when available
Most annotations can have content associated (e.g. a comment
on a highlighted text, the comment inside a post-it like note).
We dislpay that content in the sidebar, as it is saves times to
the user, and it is convenient to have there for revision.
That is, without the user going annotation after annotation to
review each content.

Fixes #386