pdf: support 'de facto' tooltip feature
Most pdf readers implement a tooltip feature by
showing the string content of 'TU' field of a
widget annotation that is not linked to any
form field.

Normally, widget annotations carry a reference to a
form field which are used together to implement the
different form widgets. But, the PDF spec does not
forbid standalone (i.e. not linked to any form field)
widget annotations, and the fact is they're been used
by most pdf readers to show a tooltip when the area
of that AnnotWidget is hovered.

According to issue reported files, 'pdfTeX' seems to
be one of the pdf producers incorporating this tooltip
feature.

The support for this feature has been recently been
added to poppler-glib with following api:

gchar   *poppler_annot_get_alternate_name (PopplerAnnot *annot)
gboolean poppler_annot_set_alternate_name (PopplerAnnot *annot,
                                           const gchar  *name)

so this commit uses only the "read" part of that api
to show tooltips on pdf's that carry them. For that
we created a new EvAnnotationTooltip type. This new
annot type is not shown on the annotations sidebar
list as part of the rest of annotations, instead it's
just been used to show the tooltips when the annot
area is hovered.

Evince issue #842

Poppler issue:
https://gitlab.freedesktop.org/poppler/poppler/issues/34