EvToolbar: set menu button to not focus-on-click
Gtk3 has just merged a commit that makes all
GtkMenuButton's to not focus-on-click, because
as a general rule, all buttons that launch a
menu should not grab focus on click, because
otherwise when the menu is closed the focus goes
back to the button instead of the previously
focused widget, which is the one the user was
interacting with.

Because the Gtk fix will be in next 3.24
release and Evince requires >= 3.22 let's
apply this easy fix to Evince now.

Fixes #1339