view: Fix "can-have-popup" prop when creating markup annotations
Highlight annotations created directly from a selection do not
display a popup when clicked. This happens because
ev_view_create_annotation_real neither creates a window for the
newly created annotation, nor sets the "can-have-popup" property
for the annotation. Consequently, when the annotation is clicked,
ev_view_handle_annotation never creates a window for the annotation.

Fixes #1373