libview: use annotation color for the close button
Currently, the close button of annotations takes the color from
the theme on hover, which means it appears white/black on the
colorful headerbar.

This commit stop using the internal close button icon and now
uses GTK window-close-symbolic instead, which means it can now
be colored as well. The deprecated GTK object GtkStyleProperties
has been replaced by GtkCssProvider. Finally, the color of the
button is now the same as annotation's one and the close icon
takes the most readable color between white and black.