cut-n-paste: Use async calls in file manager
Avoid blocking the UI if `dzl_file_manage_show` is called from
the main thread if there were any issue with dbus (no active
or slow).

`dzl_file_manage_show` is a sync call in three systems (Win32,
OSX, and Linux), and ideally, we could have it async for all
platforms, but this is a start, albeit an ugly one (ugly
because of the mix of sync and async calls between platform
ifdefs.