dvi: Fix some math fonts when using virtual fonts
Virtual fonts that have three or more mappings, e.g. zptmcm7y.vf and
mdugmri7m.vf, do not display properly, giving warnings like:

/home/me/test/evince-mdugmri7m.dvi[0]: Error: font 1 is not defined
/home/me/test/evince-mdugmri7m.dvi[0]: Warning: mdugmri7m: vf macro
  had errors

This commit fixes the logic by which vf_load_font() creates a
DviFont* containing a linked list of DviFontRef*s. In an example with
3 fonts it contained references to font 0 and font 2, but not font
1. The pointer to font 1 was being overwritten by font 2 because the
DviFont* last was not set to the current ref at the end of each
iteration of the main loop. This commit fixes the problem by setting
last to ref.

Fixes #53