dvi: Minus sign doesn't appear with mathdesign fonts
The minus sign does not appear in the dvi file that results from
latex'ing:

  \documentclass{article}
  \usepackage[mdugm]{mathdesign}
  \begin{document}
  $x-y$
  \end{document}

Evince gives the following warning when rendering the dvi file:

  Warning: requested character 0 does not exist in `mdugmr7y'.

The offset of the DviFontChar for the minus sign is 0, so that
glyph_present() reports the character as being absent, when it is in
fact present. This patch fixes the problem by only checking the
DviFontChar pointer is not null, removing the test that the content
of DviFontChar is 0.

Fixes #1477