window: Remove fullscreen overlay
The current fullscreen mode does not work correctly in the
scenesario where there is an additional toolbar (e.g.
a search bar and the annotation bar). In that situation,
the overlay makes the headerbar be on top of the additional
toolbar.

Something similar occurs with the sidebar, which the top
left part becomes hidden when the headerbar is displayed
after some user interaction.

On Wayland, the screen shows two similar toolbars.

All in all, the short term solution is to remove the overlay,
introducing a regression as the toolbar is now always visible
Showing chrome  is not what the user may expect on a full
screen session. However, this removal makes Evince usable again
in fullscreen mode.

Closes #1442, #336