libview: forgo setting device offset on page surfaces
I believe that passing unscaled offset coordinates to cairo_set_source_surface
during a process of drawing a page _effectively_ does the same transformation.

Also this fixes link preview popups when a link is on the currently viewed
page, because the solution introduced in commit caad329a did not account for
nonzero device offsets. Without this patch any preview of a link referring
to the current page would go completely black when the page is scrolled at
least popup's height pixels down.