window: Show password dialog again
Show password dialog when trying to open a document for which we've
cancelled a password dialog before in recent view.

This is achieved by adding a signal "cancelled" to EvPasswordView since
only successfull entering of password was signaled before by "unlock" signal.

We also need to unregister current document with EvApplication and clear its uri
there to be able to open other documents in current window.