libdocument: get backends directory from environment
When bundling evince library, the directory of the backends is copied in
another directory than the installation directory.

For example, cx_freeze is a popular Python package for creating
standalone executables from Python scripts. For applications using
Evince as a library, cx_freeze copies the evince dlls in the main
directory with the executable and the backends are copied to this
directory. So the directory is relative to the directory where the
executable is installed and not from the package installation
directory of evince (because it is not really installed).

Therefore, to help the use case where Evince is bundled as a library,
and not installes, we need a way to specify dynamically the path of this
new directory.

See MR !327 for more discussion.