shell: Explicitly set title in the header bar as a fallback
Before migrating to HdyHeaderBar, gtk_window_set_title was enough
to set the title in the window and GtkHeaderBar because GTK internally
synced both. That bevahiour does not occur with HdyHeaderBar, thus
HdyHeaderBar was updating the title/subtitle only when the document
and the filename were set, leaving the title in blank otherwise
That is, when both title_header and subtitle were set, and did nothing
otherwise.

We now check than in any other situation, we fallback to the title,
which would take a proper value if there is no document, or only we
have its file name.

Fixes #1583