Outline: be able to collapse/expand all entries
This commit adds three new context menu actions
to the Outline tree view:

1) Collapse all tree
2) Expand all tree
3) Expand all under this element

Number 3) is like 2) but restricted to the
children of the currently selected row. We
insensitivize this action in the menu when
we don't have any grandchildren.

All three actions properly update the metadata
keys "index-expand" and "index-collapse" so
the state of the tree view is reinstated in
future Evince invocations.

Fixes #736