properties: Show None when missing creation/modification date
According to the PDF spec, the creation date and modification date are
optional (for pdf documents). We can safely assume that a file will not
to be modified exactly at 31 Dec 1969 23:59:59 UTC, so each creation and
modification date equal to -1 is considered as NULL for the pointers to
the other attributes (None is displayed).

The if statements could be used in the future when using GDateDime instead
of GTime (a pointer could help, or optional type), and it works for
any type of document, not only PDF.

Closes: #1549

Signed-off-by: Philippe Volte--Vieira <philippeisvv@hotmail.fr>