shell, previewer: use box-shadow to draw the shadow when possible
This is a major performance boost since the old border-image way still
triggers fallback drawing code path and uses cairo to draw the border
every time. In my 4K desktop, drawing frame by the old way makes the
FPS drop to 20 which is significant performance issue.

The border-image is still not fully removed, since it's still in use in
EvRecentView, and will be until the GTK4 port.

In addition, the default spacing in EvView needs to be increased, since
now the pixels used to draw the rendered pages do not count as part of
them.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>
Co-authored-by: Pablo Correa Gomez <ablocorrea@hotmail.com>