build: Remove cruft
The shell code is not an installed shared library, so it has no use for
a reference documentation.  Since also this documentation hasn't been
updated in forever, and is the only reason shell/meson.build creates a
'libshell' shared library (which the next commit is going to remove),
let's remove this extraneous directory.