shell: Use tabular figures for zoom level and page numbers
Without tabular figures, the decimal point and the leftmost
digit move around when zooming in and out (following the
proportional size of each digit).

With tabular figures, each figure uses the same space and
when zooming in and out (or switching pages), the numbers
stay in place.

See https://gitlab.gnome.org/GNOME/evince/-/merge_requests/414