shell: allow EvView to correctly receive focus
This fixes bad behaviour described in following comment:
https://gitlab.gnome.org/GNOME/evince/-/issues/704#note_1350123

which is than when TAB key switching focus to the document view
it does not get it the first time, you have to play some UP and
DOWN arrow till it gets it (as described in previous comment).

The reason is that GTK does not pass the focus directly to EvView
but to the GtkScrolledWindow that contains it. But we really want
the focus to end up in the EvView, so we manually do that so the
focus TAB chain performs as expected.

Part of issue #704