EvRecentView: fix crash when thumbnail extraction takes too long
When opening a large file for the first time from the recent
view, evince launches a background process to extract and save
the thumbnail of the first page, this process can take several
seconds for big documents (like comic files).

So it can happen that when it finished extracting the thumbnail
the EvRecentView has already been destroyed (to show document
view) and so when the "save thumbnail" code tries to access
the recent view it crashes.

Fixed by refing the view when attaching it to async process
and unrefing it in its corresponding free_data callback.

Fixes issue #965