comics: Speed up fetching page sizes
Significantly speed up fetching page sizes by keeping track of which
order the files are inside the archive.

Before this patch, the code would always open the archive again, loop
until we got to the page of interest, decode enough of it to fetch the
sizes, and close the archive.

As libarchive cannot reset, but can skip forward pretty fast, don't
reset the archive each time, and use the old instance as long as the
next page of interest is placed later in the archive.

On a 200 pages CB7 file, this cut down the page size detection from 2
minutes 15 seconds to 3 seconds.