build: Require poppler 22.02.0 for PDF rendering
Even though Evince can be built with older versions of Poppler,
this plays against users, who depends on distributions to update
the version of Poppler to build Evince. In other words, we are
defering the decision to distributions to determine which version
is better for the users.

However, as we follow closer the development of the libraries
we depend on (at least the most important ones), we are likely be
in a better position to say which version is best for the users
and that is stable.

As of today, poppler is very stable, and the development is
incremental, and they try hard to avoid regressions, as much as we
do.

My motivation to bump the requirement is that I saw users stating
about some bugs or features missing in Evince that has been supported
for more than a year, but Evince could not take advantage of because
the Poppler version was old. For example, searching words in multiple
lines. Like this, there has been several fixes in Poppler that were
originally reported in Evince. If we do not "force" to update Poppler,
it will take longer for the users to see these fixes, and probably
blame Evince because of that.