shell: Reset window title when cancelling password dialog
Currently we do not reset the Header bar title when the user
backs off from the password dialog in the recent documents view.

When a password protected file is opened, and its password
dialog cancelled, the header bar title remain as
"Password Required", and the filename is not updated, despite
being back to the recent documents view.

To fix this, we set the proper window title (Recent type) when
we cancel the password dialog, and set the title of the header
bar when the window type is Recent (done whenever the window
title is updated).

Fixes #931