shell: Avoid using gtk_widget_show_all on ev-zoom-action
Facilitate the transition to gtk4 by avoid using a call that
does not exist, and the same can be achieved using composite
template.

Do not set visibility in main container, as it will be shown
or hidden in the application depending if the mode is Recent
View, Normal, or Fullscreen.