shell: Use composite template for EvToolbar
This adapts to the most recent used pattern and simplifies
the code.

It was required to modify EvZoomAction to fit the template model.
EvZoomAction needs the model (EvDocumentModel) to be fully-functional.
The said model is not available at the widget creation time when using
templates because the model is only available programmatically. Therefore,
we partially revert commit 55b21fe75eb910992af06a23fdd11d3073c06f6e
by using again ev_zoom_action_set_model. In the process, the "document-model"
property is also removed, as it is not needed (never emmited, no consumers).
The G_PARAM_CONSTRUCT_ONLY condition of the property is partially kept by
making sure that we throw a critical if the model is already set.