use default color when annotation has no color
Annotations may come with a NULL color, use
a default color there instead of letting
undefined GdkRGBA to crash the application.

We take the chance to share this default color
to be the same we already use when creating
an annotation from Evince.

We also take the chance to replace deprecated
function ev_annotation_set_color() with
ev_annotation_set_rgba()

Fixes crash in #1800