backends: Handle labels as optional on markup annotations
Label is an optional property of a markup annotation, so when Evince
reads markup annotations which have no label it should not print
critical messages.

Therefore, fon't set label property with a NULL value in
ev_annot_from_poppler_annot() function of ev-poppler.c because
otherwise a critical message will be printed to the console.

Fixes #1814