Go to requested page for links to external PDF documents
Links to external documents in PDF (GoToR actions) specify either a page
number or a named destination in the target. Evince ignored the page number
and external PDFs were opened in the first or last visited page. Jumping to
the appropriate page number was considered in ev_spawn, but only for the
destination type EV_LINK_DEST_TYPE_PAGE (which should not occur in normal
conditions) and not for EV_LINK_DEST_TYPE_XYZ and others.