sidebar-links: tweak 'Expand all under this element' item sensitivity
Two changes:

1) Same as with 'Print section' item, only enable it when
clicking over the selected path. This avoids confusion as
the user may previously doubt if the action was for the
selected path or the path under the mouse pointer.

2) Revert 'enabling only when the selected path has
grandchildren'. Avoids confusion as the user may doubt
why the action is disabled when there's an arrow to
expand the path. Now we only check if it has any child
to enable the action.