ev-view: remove non-needed call to deprecated gtk_style_context_set_background
This was introduced in 17859243df98bcd9698c67081f2443a08e60caeb with
the rationale:

Even if we now render a background in _draw() with
gtk_render_background, using gtk_style_context_set_background() in
realize makes the backing GdkWindow not being completely transparent,
which enables GTK to apply optimizations during the paint cycle.
The result is, especially in clutter-gtk applications such as Documents,
scrolling in a document page is much smoother.

https://bugzilla.gnome.org/show_bug.cgi?id=692988

Clutter was completely archived in GNOME 42 https://blogs.gnome.org/clutter/
so the rationale does no longer hold. Since the call is deprecated, just
remove it, as it is unncessary.