ev-attachment: deprecate GTime related functions for GDateTime variants
8257a0555b0666fa38d7ec2692f2ba6f6a182de0 added some IGNORE_DEPRECATIONS
markers, but did not fix the underlying issue in ev-attachment, the
use of GTime. This commit provides GDateTime variant alternatives for
all the functions that still use GTime. This allows to
straight-forward remove the deprecated functions after the GTK4 move.

The GDateTime mtime and ctime are added as fields in the EvAttachmentPrivate
structure, but not presented as properties in EvAttachment. Those properties
are completely used in Evince's codebase, and likely unnecessary.