ev-debug,ev-jobs: replace EvProfileSection for a boolean
There's only job profiling, so simplify the code that makes use of
it. This might have been a nice interface design in the beginning,
but it's been 15 years since 597433bc874ff2cf5cb9f9de4e053db47837e042,
and it has not been put into use. So clean it up.