ev-jobs: do not consider profiling when operation is cancelled
The timings in such case are irrelevant, since it depends on user
action, and not on the system status. In the current implementation with
timers, this has the consequence that the timer is never stopped, but
that is going to change in following commits, solving this problem.

As a consequece, ev_profiler_start becomes unused. It marks the start of
the profiling, but ev_profiler_stop is never called with that same string,
so this is a timer that will never be stopped.