shell: remove 'annotate-selected-text' action
which was used in the EvView context menu, and
use instead 'highlight-text' action, which
provides the same functionality (highlighting
selected text) but with the added bonus of
showing the keybinding (Ctrl+H) in the context
menu.

In case the action was disabled for the context
menu we need to enable it back again when
closing the menu and returning to normal view,
we do that in the menu "hide" signal handler.

Closes #1919