ci: add check for code style respecting the repository .editorconfig
It currently runs with checks for indentation disabled. The main reason
is that the code contains innumerable situations where spaces are used
for indentation instead of tabs. Fixing those require extensive manual
and repetitive labour. And having this is probably better than nothing.

Additionally, the cut-n-paste is also ignored, since otherwise it wouldn't
be "cut and paste".