ev-application: remove loading and saving of custom accels file
This is an undocumented feature, since the ability to change shortcuts
was removed somewhere in GTK 3.10, according to the commit that
updated the documentation: 54f43719c658a0bff3946bf2be2a7a76d679773a

The shortcuts documentation was later completely removed in
a050ee73d7a55b489c085ff81e654c656043e1fa since the help-overlay in the
shell takes care of it.

Another reason to drop this is the lack of first-class support in
GTK4:

"There is no direct replacement for loading and saving accelerators
with GtkAccelMap. But since GtkShortcutController implements
GListModel and both GtkShortcutTrigger and GtkShortcutAction can be
serialized to strings, it is relatively easy to implement saving and
loading yourself."

So in GTK4 it's possible, but requires further development. Which is
likely of very little use for an undocumented feature introduced 15
years ago as a copy-paste in 83a7a11ba109b8aee5cdc047dec8b93967ccfc0e

Relates !373