Remove evince.ico formerly used for windows resources
Introduced in 45467f0d8a4dacfe4209fba0a605e86b2b2e9b80 and never
touched since, the icon is a way to embed the icon into the binary in
a windows-compliant way. At the time, there was code consuming the
ICON macro, like "cut-n-paste/smclient". That code does not exist
anymore and renders this file and the logic around it unnecessary.