build: Check for NEWS and appdata updates for new releases
At least 4 different bugs were filed against the Flathub version of
evince because Flatpak prominently shows the appdata version in the UI,
making users think that there was a bug in the packaging, or that the
package had not been updated.

Avoid this problem in the future by failing the build if the NEWS or
appdata files aren't updated on release.