ci: Let the lint job be lenient
Some MR may fail because the coding style of some files do not
follow minimal coding style, regardless whether those lines are
actually part of the MR itself.

Hence, let start with a lenient lint, and in the future re-assess
whether to keep it lenient or switching it strict failure.