data: Remove unused icons
There are no icons being used in libview, there is no longer a need
to install the icons as files. Instead, the icons requested by the
application can be shipped as resources, which is the recommended way
to do it¹. This is a breaking change for applications that might be
using the icons, but setting up icons as resources is simple.

¹ https://developer.gnome.org/documentation/tutorials/themed-icons.html