ev-document-find: remove all but one find API
This was not possible to do when the "extended" and "options" variants
where introduced due to API stability. However, the compatibility difference
is quite limited, with the input having an EvFindOptions parameter, and
the return value being EvFindRectangle. Since EvFindRectangle is simply
an EvRectagle with a few more fields, this is probably trivial for consumers
to adapt