shell: remove unused EV_TOOLBAR_MODE_FULLSCREEN
Since 98f9785f41376b4fabeda340b75f629267757f43, the toolbar has the
same behavior in regular as in FULLSCREEN mode. In addition, it's not
used anywhere in the code since
e8b1293236d62359f282af40f9a4bd231b8f1cdb. Even if the functionality in
that commit were to be re-introduced, it would not make sense to have
different macros for what is effectively the same state