libview: rename EvJobRender to EvJobRenderCairo
For the porting to GTK4, it is necessary to have a renderer based on
GdkTexture. We are not removing completely the cairo renderer, yet
in case we or someone else need it as fallback. We expect to have
both EvJobRenderCairo and EvJobRenderTexture. Temporarily we are
deprecating the cairo renderer, so we do not forget about it. Later
on we can decide whether to keep it as it is and remove the guards, or
remove the cairo renderer for good.

In the process, removed an unused EvJobRender pointer in EvView