libview: rename EvJobThumbnail to EvJobThumbnailCairo
For the porting to GTK4, it is necessary to have a thumbnail based on
a GdkTexture. However, it is also helpful to retain the older cairo
renderer to do the transition at different paces. So instead of
replacing the job with a one based on a textures, split it in two:
EvJobThumbnailCairo and EvJobThumbnailTexture.

Temporarily we are deprecating the cairo renderer, so we do not forget
about it. Later on we can decide whether to keep it as it is and remove
the guards, or remove the cairo renderer for good.