shell: don't show the progress for loading fonts in the document
We have agreed that this process should not take very long in PDF
files, that are the only documents that implement the interface. So
remove the feedback, as according to HIG should not be provided for
things that take less than 3 seconds.

In addition, there's some mention in the code to the "lock". But that
is unnecessary. The PopplerFontsIter where the fonts reside is a
transfer-full structure owned by the PDF backend. So when iterating
it, we're not iterating the document itself, but our own created
structure, which should be possible to access without a lock once the
job has finished.