libview: make EvJobFonts run as a thread, and remove updated signal
The signal is unnecessary since the previous commit, as it's now no
longer used. This makes the EvJobFind the only remaining job to run in
the main loop. It has the benefit of not having to re-run in case the
document was locked, and frees the main loop from direct document
accesses.