Activate search-outline mode from text entry
Right now the search-outline mode can only be
activated through the context menu action.

Some people miss the more practical way of entering
some text and have it search the Outline, that's
what this MR implements, with the exception when
the entered text is just a number or a roman numeral,
in that case we go directly to that page (as expected).

Fixes #1991