libmisc: Fix regression after porting to g_clear_signal_handler
Fix regression introduced in 7c28e56392, in which there
should only clear when doc_model has a value (because
doc_model could be NULL, and handlers might not, and still
be a valid situation).

The error can be reproduced whenever an action is taken, e.g.
in evince-previewer, by clicking to zoom in or zoom out.