Improve support for Adobe PDF open parameters
This commit adds partial support for Adobe PDF open parameters[1],
specifically we added the "page" and "nameddest" parameters,
which are similar to Evince's --named-dest and --page-label
commandline options respectively.

Some examples of the new parameters supported:
  - file:///home/user/Documents/sample.pdf#page=12
  - https://arxiv.org/pdf/2310.01425#nameddest=section.3
  - sample.pdf#Page=5
  - sample.pdf#NamedDest=Chapter4

Evince previously only supported a number-only parameter
for opening a page upon loading the document, eg:
  - sample.pdf#5
  - https://arxiv.org/pdf/2310.01425#7

This support is kept.

Evince already supported passing a file URI in addition
to a filename, but the number-only page parameter didn't
work when using it in a URI (as described in #2044)
nor did it work when opening a pdf URI from within an
existant pdf (see testcase pdf in here[2]).

So, this commit fixes those issues and also updates
the Evince man page to:
 - Inform of new page and nameddest open parameters.
 - Inform that a file URI can also be passed.
 - Fix dead links.

[1] https://helpx.adobe.com/acrobat/kb/link-html-pdf-page-acrobat.html
[2] https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/222#note_2197697

Fixes #2044